src/test/javascript/spec/app/entities/accountService/key-set/key-set.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { KeySetService } from 'app/entities/accountService/key-set/key-set.service';
import { IKeySet, KeySet } from 'app/shared/model/accountService/key-set.model';
import { MarketE } from 'app/shared/model/enumerations/market-e.model';

describe('Service Tests', () => {
  describe('KeySet Service', () => {
    let injector: TestBed;
    let service: KeySetService;
    let httpMock: HttpTestingController;
    let elemDefault: IKeySet;
    let expectedResult: IKeySet | IKeySet[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(KeySetService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new KeySet('ID', MarketE.SAMPLE_ENUM, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find('123').subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a KeySet', () => {
        const returnedFromService = Object.assign(
          {
            id: 'ID',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new KeySet()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a KeySet', () => {
        const returnedFromService = Object.assign(
          {
            market: 'BBBBBB',
            apiKey: 'BBBBBB',
            apiSecret: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of KeySet', () => {
        const returnedFromService = Object.assign(
          {
            market: 'BBBBBB',
            apiKey: 'BBBBBB',
            apiSecret: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a KeySet', () => {
        service.delete('123').subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
